pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rm-notebooklm"
version = "0.1.0"
description = "Sync reMarkable handwritten notes to NotebookLM/Gemini and return AI responses as PDFs"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["remarkable", "notebooklm", "ocr", "gemini", "pdf", "e-ink"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # reMarkable Cloud API (sync15 support — do NOT use rmapy)
    "rm-api==1.1.1",
    # .rm binary parsing (v6 format, firmware 3.x+)
    "rmscene==0.7.0",
    # Config from env / .env files
    "pydantic-settings>=2.0",
    # CLI
    "typer>=0.12",
    "rich>=13.0",
    # PDF generation (reMarkable 2 dimensions)
    "fpdf2>=2.7",
    # Image preprocessing for OCR
    "Pillow>=10.0",
    # Retry + circuit breaker
    "tenacity>=8.0",
    "pybreaker>=1.0",
    # Structured logging
    "structlog>=24.0",
    # Google Drive / NotebookLM Enterprise
    "google-api-python-client==2.189.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2",
    # Google Cloud Storage (Path C)
    "google-cloud-storage>=2.0",
    # Gemini SDK (OCR + grounding, Path C)
    "google-generativeai>=0.8",
    # HTTP client
    "httpx>=0.27",
    "requests>=2.31",
]

[project.optional-dependencies]
openai = [
    "openai>=1.30",
]
claude = [
    "anthropic>=0.30",
]
path-a = [
    # Unofficial notebooklm-py — cookie-based, fragile, prototyping only
    "notebooklm-py[browser]==0.3.0",
]
dev = [
    # Testing
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "pytest-recording>=0.13",   # VCR cassette recording
    "vcrpy>=6.0",
    "responses>=0.25",          # requests mock
    # OCR quality metrics
    "jiwer>=3.0",
    # PDF verification
    "pypdf>=4.0",
    # Linting + formatting
    "ruff>=0.4",
    # Type checking
    "mypy>=1.10",
    "types-requests",
    # All optional deps for testing all paths
    "openai>=1.30",
    "anthropic>=0.30",
]

[project.scripts]
rm-notebooklm = "rm_notebooklm.cli:app"

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/remarkable-notebooklm"
Issues = "https://github.com/YOUR_USERNAME/remarkable-notebooklm/issues"

# ---------------------------------------------------------------------------
# Hatch build config
# ---------------------------------------------------------------------------
[tool.hatch.build.targets.wheel]
packages = ["src/rm_notebooklm"]

# ---------------------------------------------------------------------------
# Ruff — linting + formatting
# ---------------------------------------------------------------------------
[tool.ruff]
src = ["src"]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "ANN", # flake8-annotations (light — only missing return types flagged)
]
ignore = [
    "ANN401", # allow Any
    "TC001", # don't require TYPE_CHECKING for first-party imports
    "TC002", # don't require TYPE_CHECKING for third-party imports
    "TC003", # don't require TYPE_CHECKING for stdlib imports (Path is used frequently at runtime)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "ANN001", # test helpers and fixtures don't need arg annotations
    "ANN201", # test functions don't need return annotations
]

[tool.ruff.lint.isort]
known-first-party = ["rm_notebooklm"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ---------------------------------------------------------------------------
# Mypy — type checking
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
exclude = ["tests/"]

[[tool.mypy.overrides]]
module = [
    "rmscene.*",
    "fpdf.*",
    "pybreaker.*",
    "vcrpy.*",
    "responses.*",
    "jiwer.*",
    "googleapiclient.*",
    "google.auth.*",
    "google.oauth2.*",
    "google.cloud.*",
    "google.generativeai.*",
    "notebooklm.*",
]
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "ocr: requires OCR provider API key",
    "e2e: requires real reMarkable + Google credentials",
    "slow: takes >10s",
]
addopts = [
    "--strict-markers",
    "--tb=short",
    "-q",
]

[tool.pytest-recording]
default_vcr_cassette_name = "{function}"
cassette_library_dir = "tests/fixtures/cassettes"
record_mode = "none"

# ---------------------------------------------------------------------------
# Coverage
# ---------------------------------------------------------------------------
[tool.coverage.run]
source = ["src/rm_notebooklm"]
omit = ["tests/*", "scripts/*"]

[tool.coverage.report]
show_missing = true
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
]
